Reorganizing the folder structure
